Fix redirect= order handling
The redirect= modifier should be checked _after_ mechanisms, regardless
of where it appears in the SPF record.

From https://tools.ietf.org/html/rfc7208#section-6.1":

  If all mechanisms fail to match, and a "redirect" modifier is present,
  then processing proceeds as follows:

  [...]

  For clarity, any "redirect" modifier SHOULD appear as the very last
  term in a record.  Any "redirect" modifier MUST be ignored if there is
  an "all" mechanism anywhere in the record.

To implement this, the library meant to move any redirect= modifier to
the end of the record. However, due to a typo, this was not working.  To
make it worse, there were no tests covering redirects :(

The impact of this bug was that mechanisms that came after redirect=
modifiers were ignored. This should be quite rare in practice, but it's
possible.

This patch fixes the reordering (by fixing the typo), and more
importantly adds tests for various redirect= cases.