Use static errors in most cases
In order to reduce the potential attack vector, and to make tests more
stringent, this patch makes the errors generated by the library static.

The tests are extended to check for the particular errors as well.

Some errors from the standard library still may be propagated upwards;
those will be addressed later if needed.

Note the errors are not exported, and not a part of the API; this is
just for security and testability purposes.