tests: Isolate DNS overrides
The DNS lookup is overridden for testing purposes. In the tests today
we have a global set of arrays to do this, which is difficult to
clear in between tests, and awkward to manage.

This is ok for the test suite as it is today, but with the extended
automated tests we want to be able to clear the slate easier, so this
patch moves the overrides to their own structure and file.

The existing tests are adjusted to reset the overrides on each test.