a, mx: Support dual CIDR length masks
For A and MX mechanisms, the masks are specified with a special syntax,
allowing no/IPv4/IPv6/both masks.

The form is "[/<mask4>][//<mask6>]". Note that both are optional, and if
only mask6 is given, it still needs to be prefixed by "//".

The current implementation doesn't handle this well, so this patch
extends the mask management significantly to support the masks properly.

Found by the standard test suite.