Allow overriding the maximum lookup limit
This patch allows the user to override the maximum lookup limit, which
today is hard-coded to 10 (as per the RFC).

This enables non-standard behaviour, which can be useful under some
circumstances (like testing tools).

The change should be fairly future-proof since the Option type is
opaque, although it is exported for the callers' convenience.

Amended-by: Alberto Bertogli <albertito@blitiri.com.ar>
  Wrote commit message, added tests, did some naming and style
  adjustments.