Allow passing a context.Context to the resolver functions
In some cases it can be useful to propagate a context through the SPF
checks, in particular to make sure that the resolver functions honour
context cancellations.

This patch adds a new WithContext option, that lets user give a context
for each operation, which will be passed along to resolver functions.