Allow using a custom DNS resolver
In some cases, like for integration testing or when running in special
environments, the users may need to force the library to use a non-default
DNS resolver.

This patch adds an option for that, via a new WithResolver option.

Thanks to Thierry Fournier <thierry.fournier@ozon.io> for the bug report
and alternative patch.