Use net.IPMask to represent masks internally
Currently we we internally use ints to represent IP masks, and then
convert them at checking time.

This carries unnecessary complexity, as we need to check the validity of
the masks in a couple of different places.

This patch simplifies the handling by moving to representing masks with
net.IPMask, which in turn simplifies a couple of helper functions.

Some tests are added to validate corner cases continue to work as
expected.