tests: Share DNS test helpers between go and fuzz tests
We need to update the fuzz tests to use the new DNS helpers;
unfortunately, the fuzzing infrastructure ignores _test.go files and
can't be reasonably made to include them.

To avoid duplicating or complicating things more, this patch moves the
dns test helpers to its own package, renaming functions and fields
accordingly.

In the future, once fuzz support is part of the standard testing
infrastructure and we can include _test.go files, we can rename the file
again and exclude it completely from the default build.