Export error definitions
This patch exports the error definitions, to enable users to
programmaticaly distinguish between the different errors, and handle
them accordingly.

This, however, has to be done with care, as errors can evolve in the
future (new ones added, reasons changed, etc.).

That said, current ones have been stable for a while, so not a lot of
churn is expected.

In the future if there are enough use cases, we may introduce some
additional types and wrapping to allow users to group them. But for now,
that complexity is not justified, and would introduce API limitations
that would be harder to undo later.