Add DNS "void lookup" count checks
RFC7204 section 4.6.4 states:

> SPF implementations SHOULD limit "void lookups" to two. An
> implementation MAY choose to make such a limit configurable. In this
> case, a default of two is RECOMMENDED.  Exceeding the limit produces a
> "permerror" result.

A DNS "void lookup" is one that returns either an empty answer, or
NXDOMAIN.

Limiting the number of void lookups when doing SPF checks helps
reduce/prevent DNS amplification attacks (RFC7204 section
11.1, 4th bullet).

This patch introduces a void lookup check, with a default limit of 2 as
per the RFC recommendation, and an option to override it.

For reference, such a check already exists in pyspf, and it seems to
also be enforced by some large providers. An exploration of the top-1000
sites on the Alexa ranking showed all domains to be okay except for 3,
which were misconfigured for other reasons and would have issues
regardless.

See https://github.com/albertito/spf/pull/2 for more background and
discussion.

Amended-by: Alberto Bertogli <albertito@blitiri.com.ar>
  Edited commit message; minor comment indentation changes; added
  tracing; removed "skip" from a standard test, now that this limit is
  enforced; unified individual commits into a single one for history
  readability.