Fix DNS lookup limits in PTR field processing
When processing PTR fields, we have to do DNS A/AAAA lookups. The RFC
defines a special logic for those, different than the other lookups
for the other mechanisms.

In particular, we must only consider the first 10 A/AAAA resource records,
and drop the rest (instead of erroring out); and those don't get counted
for the general limit.

This patch fixes that issue by adjusting the checks to match the RFC.