Fix DNS lookup limit off-by-one error in the count
As per RFC, we should not count the initial DNS lookup for the purposes
of the lookup limit, but the code currently does.

This can result in over-limiting by 1 lookup.

This patch fixes the problem by incrementing the counts on their
respective mechanism and modifier. That behaviour also matches the RFC
wording, so it is easier to follow.

https://datatracker.ietf.org/doc/html/rfc7208#section-4.6.4

Thanks to Anh Do <doducanh2710@gmail.com> for reporting this issue in
https://github.com/albertito/spf/issues/3.