Fix DNS lookup limits in MX field processing
The RFC text for what to do with MX fields in the context of DNS lookup
limits is:

> When evaluating the "mx" mechanism, the number of "MX" resource
> records queried is included in the overall limit of 10 mechanisms/
> modifiers that cause DNS lookups as described above.  In addition to
> that limit, the evaluation of each "MX" record MUST NOT result in
> querying more than 10 address records -- either "A" or "AAAA"
> resource records.  If this limit is exceeded, the "mx" mechanism MUST
> produce a "permerror" result.

Today, we count each individual MX record (which triggers a DNS lookup)
against the total term counter. That is contrary to the RFC, and what
other libraries do.

The correct behaviour, which this patch implements, is to count each
"mx" term as 1, regardless of how many MX records it produces. This
enables some level of DNS amplification, but because there is a second
limit of 10 MX records per term, it is bounded.

This makes our behaviour match RFC, and also other popular
implementations.