In MX lookups, continue on DNS NXDOMAIN errors
The RFC is clear that NXDOMAIN errors should be treated as if the server
returned zero answers.

https://datatracker.ietf.org/doc/html/rfc7208#section-5

But today, when an MX record doesn't exist, we break with an error.

This was reported in https://github.com/albertito/spf/issues/5.

This patch fixes this problem by skipping MX records which return
NXDOMAIN.

Amended-by: Alberto Bertogli <albertito@blitiri.com.ar>
  Updated commit message, renamed variable for consistency.