test: Add tests for NXDOMAIN responses
The RFC is clear that NXDOMAIN errors should be treated as if the server
returned zero answers: https://datatracker.ietf.org/doc/html/rfc7208#section-5

Based on that, some specific behaviour is expected; for example on how
to handle non-existing MX records, or a non-existing redirect target.

This patch adds tests for some of those behaviours (the ones we already
handle correctly and are not otherwise covered), to prevent future
regressions.

Thanks to Matthias Schneider <ms@wck.biz> for reporting these issues in
https://github.com/albertito/spf/issues/5.