Explicitly differentiate between a permanent error and NXDOMAIN
At the top-level check, today we treat any permanent error as if it was
an NXDOMAIN error, which is special because it signals the domain
doesn't exist, and the RFC has specific behaviour expectations for it.

This is because originally Go did not provide a good way of
distinguishing the two, but that is no longer true since Go 1.13.

The consequences are that top-level and some recursive DNS permanent
errors return None instead of PermError, which is not ideal as it can be
misleading for some use cases.

This patch fixes the problem by adding code to handle NXDOMAIN
explicitly.