test: Update fuzz tests to Go's standard toolchain
The fuzz tests were written before Go added fuzzing support to its
standard toolchain, in Go 1.18.

This patch updates them to the standard toolchain.

Note that a tag must still be specified manually in order to run them,
because the minimum supported Go version (1.15) does not have fuzzing
support, and we don't want tests to fail because of it.

Once our minimum supported version is >= 1.18, we will enable them by
default.