Replace workaround to detect NXDOMAIN with actual checks
The RFC is clear that DNS errors other than RCODE 3 (NXDOMAIN) should
result in a "temperror":

> Several mechanisms rely on information fetched from the DNS.  For
> these DNS queries, except where noted, if the DNS server returns an
> error (RCODE other than 0 or 3) or the query times out, the mechanism
> stops and the topmost check_host() returns "temperror".  If the server
> returns "Name Error" (RCODE 3), then evaluation of the mechanism
> continues as if the server returned no error (RCODE 0) and zero answer
> records.

https://www.rfc-editor.org/rfc/rfc7208#section-5

Originally, Go lacked the ability to distinguish RCODE 3 (NXDOMAIN) from
other causes, so the code had to assume all DNS permanent errors were
NXDOMAIN as a (very imperfect) workaround.

That causes it to treat legitimate errors in the same way as NXDOMAIN,
which is incorrect.

Now that Go has implemented an API to detect NXDOMAIN (and we already
use it), we can replace those workarounds with proper checks, and
properly respond temperror when we should.

Thanks to Matthias Schneider <ms@wck.biz> for reporting this issue and
providing an alternative patch in
https://github.com/albertito/spf/pull/8 !